backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
lingua-language-detector==2.0.2
argostranslate==1.9.6
